Study on this and prototype
Read the material and answered the questions to the best of my ability.
Some of the material still feels like it is going way over my level of
deeper understanding at the moment, but the basic concepts gel quite
nicely. I expect as time goes on the rest of what didn't make sense
will start to seem like common sense.